refactor(store): extract element, group, relationship CRUD slices + internals
- internals.ts now owns the nanoid generator, undoSnapshot, and
  pushUndoSnapshot so any slice can mutate the workspace and push undo
  consistently.
- element-slice: addPerson, addSoftwareSystem, addContainer, addComponent,
  updateElement, updateElementLive, updateElementTechnology, deleteElement,
  deleteElements, duplicateElements.
- group-slice: addGroup, updateGroup, deleteGroup.
- relationship-slice: addRelationship, updateRelationship,
  reconnectRelationship, deleteRelationship.
- workspace.ts shrinks 842 → 467 LOC, dropping below the 500-LOC ceiling
  the simplicity audit flagged. Remaining workspace.ts content is view
  CRUD + tag/style + undo/redo + lifecycle — extracting next.

969/969 vitest passing.
